[Feature] Sort results table by headers (#58)
Related to #53

This PR adds sorting to the results table. This can be done by clicking
on the header of the table.

For narrow columns, the arrow indicating the ordering is underneath the
column title instead of next to the title - not sure how to solve that
elegantly without a minor redesign of the table :D


![image](https://github.com/user-attachments/assets/2d2f6d4f-681a-4900-b55a-67642cc3169f)